sbus: Multiplication result converted to larger type
A multiplication result that is converted to a larger type can be a sign
that the result can overflow the type converted from. Using a cast to
the target type (size_t) to avoid overflow.

+++ src/sbus/interface/sbus_iterator_readers.c
@@ -87,7 +87,7 @@ _sbus_iterator_read_basic_array(TALLOC_CTX *mem_ctx,
     switch (dbus_type) {
     case DBUS_TYPE_STRING:
     case DBUS_TYPE_OBJECT_PATH:
-        array = talloc_zero_size(mem_ctx, (count + 1) * element_size);
+        array = talloc_zero_size(mem_ctx, (size_t)(count + 1) * element_size);
         if (array == NULL) {
             ret = ENOMEM;
             goto done;
@@ -106,7 +106,7 @@ _sbus_iterator_read_basic_array(TALLOC_CTX *mem_ctx,
             goto done;
         }
 
-        array = talloc_zero_size(mem_ctx, count * element_size);
+        array = talloc_zero_size(mem_ctx, (size_t)count * element_size);
         if (array == NULL) {
             ret = ENOMEM;
             goto done;
